:art: Separate DTO objects from routing
